refactor(handlers): extract git.ts handler from intent.ts (Stage B)
handleSAPGit + backend resolver/repo-load helpers (abapGit + gCTS). 3672 tests, A1 zero-diff.
